+++ tools/checksums.py
@@ -11,6 +11,7 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 import setup
 
 sqlitevers = (
+    "3530400",
     "3530300",
     "3530200",
     "3530100",