+++ doc/changes.rst
@@ -13,9 +13,10 @@ APSW changes by version
 next
 ====
 
-:class:`Blob` implements :class:`io.IOBase` allowing use by other
-modules expecting file like objects, such as compression.
-(:issue:`623`)
+:class:`Blob` implements :class:`io.RawIOBase` allowing use by other
+modules expecting file like objects, such as compression.  Several
+methods and attributes are added, and :meth:`Blob.write` now returns
+the size written (used to be None).  (:issue:`623`)
 
 3.53.3.1
 ========